auth/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
sqlmodel==0.0.8
asyncpg==0.27.0
python-multipart==0.0.6
email-validator==2.0.0


